Move execution-related code under the resolvers module
