+++ crates/apollo-compiler/src/resolvers/execution.rs
@@ -4,11 +4,11 @@ use crate::collections::HashSet;
 use crate::collections::IndexMap;
 use crate::executable::Field;
 use crate::executable::Selection;
-use crate::execution::input_coercion::coerce_argument_values;
-use crate::execution::result_coercion::complete_value;
 use crate::introspection::resolvers::SchemaMetaField;
 use crate::parser::SourceMap;
 use crate::parser::SourceSpan;
+use crate::resolvers::input_coercion::coerce_argument_values;
+use crate::resolvers::result_coercion::complete_value;
 use crate::resolvers::MaybeAsync;
 use crate::resolvers::MaybeAsyncObject;
 use crate::resolvers::MaybeAsyncResolved;
