+++ crates/apollo-compiler/src/resolvers/input_coercion.rs
@@ -3,11 +3,11 @@ use crate::ast::Value;
 use crate::collections::HashMap;
 use crate::executable::Field;
 use crate::executable::Operation;
-use crate::execution::engine::ExecutionContext;
-use crate::execution::engine::LinkedPath;
-use crate::execution::engine::PropagateNull;
 use crate::parser::SourceMap;
 use crate::parser::SourceSpan;
+use crate::resolvers::execution::ExecutionContext;
+use crate::resolvers::execution::LinkedPath;
+use crate::resolvers::execution::PropagateNull;
 use crate::response::GraphQLError;
 use crate::response::JsonMap;
 use crate::response::JsonValue;
