+++ crates/apollo-compiler/src/resolvers/result_coercion.rs
@@ -1,11 +1,11 @@
 use crate::executable::Field;
-use crate::execution::engine::execute_selection_set;
-use crate::execution::engine::try_nullify;
-use crate::execution::engine::ExecutionContext;
-use crate::execution::engine::ExecutionMode;
-use crate::execution::engine::LinkedPath;
-use crate::execution::engine::LinkedPathElement;
-use crate::execution::engine::PropagateNull;
+use crate::resolvers::execution::execute_selection_set;
+use crate::resolvers::execution::try_nullify;
+use crate::resolvers::execution::ExecutionContext;
+use crate::resolvers::execution::ExecutionMode;
+use crate::resolvers::execution::LinkedPath;
+use crate::resolvers::execution::LinkedPathElement;
+use crate::resolvers::execution::PropagateNull;
 use crate::resolvers::AsyncObjectValue;
 use crate::resolvers::AsyncResolvedValue;
 use crate::resolvers::MaybeAsync;
